fix(factorisation): factorisation des manager et création d'un abstractmodel
