+++ sql/blog_maxime.sql
@@ -5,7 +5,7 @@ use `blog_maxime`;
 START TRANSACTION;
 
 CREATE TABLE `user` (
-                id_user INT NOT NULL AUTO_INCREMENT,
+                id INT NOT NULL AUTO_INCREMENT,
                 username VARCHAR(45) NOT NULL,
                 email VARCHAR(45) NOT NULL,
                 password VARCHAR(45) NOT NULL,
@@ -76,7 +76,7 @@ values
 
 ALTER TABLE `article` ADD CONSTRAINT id_author_fk
 FOREIGN KEY (id_author)
-REFERENCES `user` (id_user);
+REFERENCES `user` (id);
 
 ALTER TABLE `comment` ADD CONSTRAINT id_post_comment_fk
 FOREIGN KEY (post_id)
