+++ src/Models/AbstractModel.php
@@ -1,0 +1,27 @@
+<?php 
+
+namespace App\Models;
+
+abstract class AbstractModel{
+  private $_id;
+
+  public function __construct(array $data){
+    foreach ($data as $key => $value) {
+      $method = 'set'.ucfirst($key);
+      if (method_exists($this, $method)) {
+        $this->$method($value);
+      }
+    }
+  }
+  public function setId($id)
+  {
+    $id = (int) $id;
+    if ($id > 0) {
+      $this->_id = $id;
+    }
+  }
+
+  public function getId(){
+    return $this->_id;
+  }
+}
