+++ src/Models/Comment.php
@@ -5,38 +5,14 @@ namespace App\Models;
 /**
  *
  */
-class Comment extends Model
+class Comment extends AbstractModel
 {
-
-  private $_id;
   private $_post_id;
   private $_comment;
   private $_author;
   private $_comment_date;
 
-  public function __construct(array $data){
-    $this->hydrate($data);
-  }
-
-  //hydratation
-  public function hydrate(array $data){
-    foreach ($data as $key => $value) {
-      $method = 'set'.ucfirst($key);
-      if (method_exists($this, $method)) {
-        $this->$method($value);
-      }
-    }
-  }
-
   //setters
-
-  public function setId($id)
-  {
-    $id = (int) $id;
-    if ($id > 0) {
-      $this->_id = $id;
-    }
-  }
 
   public function setPostId($post_id)
   {
